Fix layer name based on clean-architecture
