--- core/build.gradle
@@ -1,93 +1,0 @@
-apply plugin: 'com.android.library'
-
-apply plugin: 'kotlin-android'
-
-apply plugin: 'kotlin-android-extensions'
-
-apply plugin: 'kotlin-kapt'
-
-apply plugin: 'dagger.hilt.android.plugin'
-
-apply from: '../dependencies.gradle'
-
-android {
-    compileSdkVersion versions.compileSdk
-
-    defaultConfig {
-        minSdkVersion versions.minSdk
-        targetSdkVersion versions.compileSdk
-        versionCode versions.versionCode
-        versionName versions.versionName
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles 'consumer-rules.pro'
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    dataBinding {
-        enabled = true
-    }
-
-    compileOptions {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: "libs", include: ["*.jar"])
-    // kotlin
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$versions.kotlin"
-
-    // architecture components
-    implementation "androidx.fragment:fragment-ktx:$versions.fragmentVersion"
-    implementation "androidx.lifecycle:lifecycle-extensions:$versions.lifecycleVersion"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:$versions.lifecycleVersion"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$versions.lifecycleVersion"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:$versions.lifecycleVersion"
-    implementation "androidx.room:room-runtime:$versions.roomVersion"
-    implementation "androidx.room:room-ktx:$versions.roomVersion"
-    implementation project(path: ':model')
-    kapt "androidx.room:room-compiler:$versions.roomVersion"
-    testImplementation "androidx.arch.core:core-testing:$versions.archCompomentVersion"
-
-    // dagger
-    implementation "com.google.dagger:hilt-android:$versions.daggerHiltAndroidVersion"
-    implementation "com.google.dagger:hilt-android-testing:$versions.daggerHiltAndroidVersion"
-    implementation "androidx.hilt:hilt-common:$versions.daggerHiltVersion"
-    implementation "androidx.hilt:hilt-lifecycle-viewmodel:$versions.daggerHiltVersion"
-    kapt "com.google.dagger:hilt-android-compiler:$versions.daggerHiltAndroidVersion"
-    kapt "androidx.hilt:hilt-compiler:$versions.daggerHiltVersion"
-    androidTestImplementation "com.google.dagger:hilt-android-testing:$versions.daggerHiltAndroidVersion"
-    kaptAndroidTest "com.google.dagger:hilt-android-compiler:$versions.daggerHiltAndroidVersion"
-
-    // coroutines
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$versions.coroutinesVersion"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$versions.coroutinesVersion"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$versions.coroutinesVersion"
-
-    // glide
-    implementation "com.github.bumptech.glide:glide:$versions.glideVersion"
-    implementation "com.github.florent37:glidepalette:$versions.glidePaletteVersion"
-    kapt "com.github.bumptech.glide:compiler:$versions.glideVersion"
-
-    // debugging
-    implementation "com.jakewharton.timber:timber:$versions.timberVersion"
-
-    // unit test
-    testImplementation "junit:junit:$versions.junitVersion"
-    testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:$versions.mockitoKotlinVersion"
-    testImplementation "org.mockito:mockito-inline:$versions.mockitoInlineVersion"
-    testImplementation "org.robolectric:robolectric:$versions.robolectricVersion"
-    androidTestImplementation "androidx.test:core:$versions.androidxTest"
-    androidTestImplementation "com.google.truth:truth:$versions.truthVersion"
-    androidTestImplementation "androidx.test.ext:junit:$versions.androidxTestJunit"
-    androidTestImplementation "com.android.support.test:runner:$versions.androidTestRunner"
-    androidTestImplementation "androidx.test.espresso:espresso-core:$versions.espressoVersion"
-}
