--- core/src/main/java/com/ddd4/core/consts/Globalconst.kt
@@ -1,5 +1,0 @@
-package com.ddd4.core.consts
-
-object Globalconst {
-        var dropIt = "DropIt"
-}
