--- core/src/main/java/com/ddd4/core/di/qualifier/ForApplication.kt
@@ -1,7 +1,0 @@
-package com.ddd4.core.di.qualifier
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class ForApplication
