--- core/src/main/java/com/ddd4/core/room/AppDataBase.kt
@@ -1,11 +1,0 @@
-package com.ddd4.core.room
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import com.ddd4.model.entity.DropitEntity
-
-@Database(entities = [DropitEntity::class], version = 1, exportSchema = true)
-abstract class AppDataBase: RoomDatabase(){
-
-    abstract fun dropItDao(): DropitDao
-}
