--- core/src/main/java/com/ddd4/core/utils/DropitGlideApp.kt
@@ -1,7 +1,0 @@
-package com.ddd4.core.utils
-
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
-
-@GlideModule
-class DropitGlideApp: AppGlideModule()
