--- core/src/main/java/com/ddd4/core/utils/ToastUtil.kt
@@ -1,7 +1,0 @@
-package com.ddd4.core.utils
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.toast(message: String, duration: Int = Toast.LENGTH_SHORT) =
-    Toast.makeText(this, message, duration).show()
