+++ data/build.gradle
@@ -1,4 +1,4 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 apply plugin: 'kotlin-android'
 
@@ -10,17 +10,17 @@ apply plugin: 'dagger.hilt.android.plugin'
 
 apply from: '../dependencies.gradle'
 
-
 android {
     compileSdkVersion versions.compileSdk
+
     defaultConfig {
-        applicationId "com.ddd4.dropit"
         minSdkVersion versions.minSdk
         targetSdkVersion versions.compileSdk
         versionCode versions.versionCode
         versionName versions.versionName
-        vectorDrawables.useSupportLibrary = true
+
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles 'consumer-rules.pro'
     }
 
     buildTypes {
@@ -30,53 +30,20 @@ android {
         }
     }
 
-    buildFeatures {
-        dataBinding = true
-    }
-
-    androidExtensions {
-        experimental = true
-    }
-
-    lintOptions {
-        abortOnError false
+    dataBinding {
+        enabled = true
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8.toString()
-    }
-
-    configurations{
-        all*.exclude group: 'com.google.guava', module: 'listenablefuture'
-    }
-
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-            returnDefaultValues = true
-        }
-    }
-
-    kapt {
-        generateStubs = true
+        sourceCompatibility = 1.8
+        targetCompatibility = 1.8
     }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-
+    implementation fileTree(dir: "libs", include: ["*.jar"])
     // kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$versions.kotlin"
-
-    // android supports
-    implementation "com.google.android.material:material:$versions.materialVersion"
-    implementation "androidx.constraintlayout:constraintlayout:$versions.constraintVersion"
 
     // architecture components
     implementation "androidx.fragment:fragment-ktx:$versions.fragmentVersion"
@@ -87,7 +54,6 @@ dependencies {
     implementation "androidx.room:room-runtime:$versions.roomVersion"
     implementation "androidx.room:room-ktx:$versions.roomVersion"
     implementation project(path: ':domain')
-    implementation project(path: ':data')
     kapt "androidx.room:room-compiler:$versions.roomVersion"
     testImplementation "androidx.arch.core:core-testing:$versions.archCompomentVersion"
 
@@ -110,9 +76,6 @@ dependencies {
     implementation "com.github.bumptech.glide:glide:$versions.glideVersion"
     implementation "com.github.florent37:glidepalette:$versions.glidePaletteVersion"
     kapt "com.github.bumptech.glide:compiler:$versions.glideVersion"
-
-    // startup
-    implementation "androidx.startup:startup-runtime:$versions.startupVersion"
 
     // debugging
     implementation "com.jakewharton.timber:timber:$versions.timberVersion"
