+++ data/src/main/java/com/ddd4/data/consts/Globalconst.kt
@@ -1,0 +1,5 @@
+package com.ddd4.data.consts
+
+object Globalconst {
+        var dropIt = "DropIt"
+}
