+++ data/src/main/java/com/ddd4/data/di/qualifier/ForActivity.kt
@@ -1,0 +1,7 @@
+package com.ddd4.data.di.qualifier
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class ForActivity
