+++ data/src/main/java/com/ddd4/data/utils/DropitGlideApp.kt
@@ -1,0 +1,7 @@
+package com.ddd4.data.utils
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class DropitGlideApp: AppGlideModule()
