+++ data/src/main/java/com/ddd4/data/utils/ToastUtil.kt
@@ -1,0 +1,7 @@
+package com.ddd4.data.utils
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.toast(message: String, duration: Int = Toast.LENGTH_SHORT) =
+    Toast.makeText(this, message, duration).show()
