--- model/src/main/java/com/ddd4/model/repository/DropitRepository.kt
@@ -1,15 +1,0 @@
-package com.ddd4.model.repository
-
-import com.ddd4.model.entity.DropitEntity
-import kotlinx.coroutines.flow.Flow
-
-interface DropitRepository {
-
-    suspend fun insert(dropitEntity: DropitEntity)
-
-    suspend fun update(dropitEntity: DropitEntity)
-
-    suspend fun delete(dropitEntity: DropitEntity)
-
-    suspend fun getAllData(): Flow<List<DropitEntity>>
-}
